.requirements.txt
google-cloud-aiplatform[adk,agent_engines]
cloudpickle
pydantic
